+++ plugins/inputs/solr/sample.conf
@@ -9,3 +9,6 @@
   ## Optional HTTP Basic Auth Credentials
   # username = "username"
   # password = "pa$$word"
+
+  ## Timeout for HTTP requests
+  # timeout = "5s"
