+++ plugins/inputs/solr/solr.go
@@ -26,111 +26,27 @@ const adminCoresPath = "/solr/admin/cores?action=STATUS&wt=json"
 
 // Solr is a plugin to read stats from one or many Solr servers
 type Solr struct {
-	Local       bool
-	Servers     []string
-	Username    string
-	Password    string
-	HTTPTimeout config.Duration
-	Cores       []string
-	client      *http.Client
-}
+	Servers     []string        `toml:"servers"`
+	Username    string          `toml:"username"`
+	Password    string          `toml:"password"`
+	HTTPTimeout config.Duration `toml:"timeout"`
+	Cores       []string        `toml:"cores"`
 
-// AdminCoresStatus is an exported type that
-// contains a response with information about Solr cores.
-type AdminCoresStatus struct {
-	Status map[string]struct {
-		Index struct {
-			SizeInBytes int64 `json:"sizeInBytes"`
-			NumDocs     int64 `json:"numDocs"`
-			MaxDoc      int64 `json:"maxDoc"`
-			DeletedDocs int64 `json:"deletedDocs"`
-		} `json:"index"`
-	} `json:"status"`
-}
-
-// MBeansData is an exported type that
-// contains a response from Solr with metrics
-type MBeansData struct {
-	Headers    ResponseHeader    `json:"responseHeader"`
-	SolrMbeans []json.RawMessage `json:"solr-mbeans"`
-}
-
-// ResponseHeader is an exported type that
-// contains a response metrics: QTime and Status
-type ResponseHeader struct {
-	QTime  int64 `json:"QTime"`
-	Status int64 `json:"status"`
-}
-
-// Core is an exported type that
-// contains Core metrics
-type Core struct {
-	Stats struct {
-		DeletedDocs int64 `json:"deletedDocs"`
-		MaxDoc      int64 `json:"maxDoc"`
-		NumDocs     int64 `json:"numDocs"`
-	} `json:"stats"`
-}
-
-// QueryHandler is an exported type that
-// contains query handler metrics
-type QueryHandler struct {
-	Stats interface{} `json:"stats"`
-}
-
-// UpdateHandler is an exported type that
-// contains update handler metrics
-type UpdateHandler struct {
-	Stats struct {
-		Adds                     int64  `json:"adds"`
-		AutocommitMaxDocs        int64  `json:"autocommit maxDocs"`
-		AutocommitMaxTime        string `json:"autocommit maxTime"`
-		Autocommits              int64  `json:"autocommits"`
-		Commits                  int64  `json:"commits"`
-		CumulativeAdds           int64  `json:"cumulative_adds"`
-		CumulativeDeletesByID    int64  `json:"cumulative_deletesById"`
-		CumulativeDeletesByQuery int64  `json:"cumulative_deletesByQuery"`
-		CumulativeErrors         int64  `json:"cumulative_errors"`
-		DeletesByID              int64  `json:"deletesById"`
-		DeletesByQuery           int64  `json:"deletesByQuery"`
-		DocsPending              int64  `json:"docsPending"`
-		Errors                   int64  `json:"errors"`
-		ExpungeDeletes           int64  `json:"expungeDeletes"`
-		Optimizes                int64  `json:"optimizes"`
-		Rollbacks                int64  `json:"rollbacks"`
-		SoftAutocommits          int64  `json:"soft autocommits"`
-	} `json:"stats"`
-}
-
-// Hitratio is an helper interface
-// so we can later on convert it to float64
-type Hitratio interface{}
-
-// Cache is an exported type that
-// contains cache metrics
-type Cache struct {
-	Stats map[string]interface{} `json:"stats"`
-}
-
-// NewSolr return a new instance of Solr
-func NewSolr() *Solr {
-	return &Solr{
-		HTTPTimeout: config.Duration(time.Second * 5),
-	}
+	client *http.Client
 }
 
 func (*Solr) SampleConfig() string {
 	return sampleConfig
 }
 
+func (s *Solr) Init() error {
+	s.client = s.createHTTPClient()
+	return nil
+}
+
 // Gather reads the stats from Solr and writes it to the
 // Accumulator.
 func (s *Solr) Gather(acc telegraf.Accumulator) error {
-	if s.client == nil {
-		client := s.createHTTPClient()
-		s.client = client
-	}
-
 	var wg sync.WaitGroup
 	wg.Add(len(s.Servers))
 
@@ -486,6 +402,8 @@ func (s *Solr) gatherData(url string, v interface{}) error {
 
 func init() {
 	inputs.Add("solr", func() telegraf.Input {
-		return NewSolr()
+		return &Solr{
+			HTTPTimeout: config.Duration(time.Second * 5),
+		}
 	})
 }
